requirements.txt
httpx==0.26.0
openai==1.10.0
notion-client==2.2.1
python-dotenv==1.0.0
